third_party/blink/web_tests/fast/forms/time/time-validity-typemismatch-expected.txt

This test aims to check for typeMismatch flag with type=time input fields

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS "" is a correct valid time string.
PASS "00:00" is a correct valid time string.
PASS "23:59" is a correct valid time string.
PASS "23:59:59" is a correct valid time string.
PASS "23:59:59.1" is a correct valid time string.
PASS "23:59:59.12" is a correct valid time string.
PASS "23:59:59.123" is a correct valid time string.
PASS " 00:00 " is an invalid time string and was sanitized.
PASS "1:23" is an invalid time string and was sanitized.
PASS "011:11" is an invalid time string and was sanitized.
PASS "ab:11" is an invalid time string and was sanitized.
PASS "-1:11" is an invalid time string and was sanitized.
PASS "24:11" is an invalid time string and was sanitized.
PASS "11" is an invalid time string and was sanitized.
PASS "11-" is an invalid time string and was sanitized.
PASS "11:-2" is an invalid time string and was sanitized.
PASS "11:60" is an invalid time string and was sanitized.
PASS "11:2b" is an invalid time string and was sanitized.
PASS "11:ab" is an invalid time string and was sanitized.
PASS "11:034" is an invalid time string and was sanitized.
PASS "23:45:" is an invalid time string and was sanitized.
PASS "23:45:6" is an invalid time string and was sanitized.
PASS "23:45:-1" is an invalid time string and was sanitized.
PASS "23:45:70" is an invalid time string and was sanitized.
PASS "23:45:zz" is an invalid time string and was sanitized.
PASS "23:45:06." is an invalid time string and was sanitized.
PASS "23:45:06.abc" is an invalid time string and was sanitized.
PASS "23:45:06.789abc" is an invalid time string and was sanitized.
PASS "23:59:59.1234567890" is an invalid time string and was sanitized.
PASS "00:00:00.0000000000" is an invalid time string and was sanitized.
PASS "invalid" is an invalid time string and was sanitized while disabled.
PASS successfullyParsed is true

TEST COMPLETE
